adds examples how to fix issue #3
each file contains some implementations of `cone`, which
is just an isl set instead of a list of indices.
These files contain checks, to verify that by 'chaining' these sets
the resulting set is the same as a specialized handwritten one.
In the triangle case this is not exactly true, since this sort of
chaning cannot produce bounds of the form `0<=j<1-i`.
The quadrilateral file also includes loopy kernel, which shows
that the generated code with these sets is as expected.